Move MetadataLoader to plugin-commons and rename to ApiVersions
